+++ dds_web/utils.py
@@ -27,18 +27,6 @@ from dds_web.database import models
 from dds_web import db, C_TZ
 
 ####################################################################################################
-# SCHEMA FIELDS #################################################################### SCHEMA FIELDS #
-####################################################################################################
-
-
-class MyDateTimeField(marshmallow.fields.DateTime):
-    def _deserialize(self, value, attr, data, **kwargs):
-        if isinstance(value, datetime.datetime):
-            return value
-        return super()._deserialize(value, attr, data, **kwargs)
-
-
-####################################################################################################
 # FUNCTIONS ############################################################################ FUNCTIONS #
 ####################################################################################################
